// MarkEd-front/src/api/models/RubricTreeNode.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
/**
 * Mirrors the jsTree node shape Mingyue's configuration page consumed.
 */
export type RubricTreeNode = {
    id: number;
    name: string;
    marks: number;
    selected: boolean;
    children?: Array<RubricTreeNode>;
};