// MarkEd-front/src/api/models/SelfAssessmentSettingSchema.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type SelfAssessmentSettingSchema = {
    assignment_id: number;
    enabled: boolean;
    use_checklist: boolean;
    use_rubric: boolean;
    use_reflection: boolean;
    deadline?: (string | null);
    needs_feedback: boolean;
    max_score: number;
};